cobrapin channel string seems to be: rgb gwr wbg brw repeat
is rgb lighting really read ABC as B->A->C? that would make this grbw grbw grbw.

Anyway this config isn't really usable except at the end of a chain, since it will eat the first channel of the next light.
And I'm not entirely sure if that would be the red or the green.....