add shows for each of the outcomes of the slots game
and play the spinner show in attract instead of the default attract show. Although, should this have a priority override?

TODO for slots is to make the row completions count towards an end show,
improve all the shows (eventually),
build in drop targets and resetting the bank,

and maybe another one or two attract shows on the grid for it to cycle
between

also testing for all the ways the mode can end (timeout, clear, low
result), maybe add time when getting a row success, and finally
find a way to integrate the mode into normal play!