add alternative control to right controlled ball gate
demo of using many diverter configs on the same coil, either with
the same or different switch/event triggers and enable/disables.

Seems like the best way to use a cbg for multiple different reasons
is to create a separate diverter instance for each reason, and let
the enable states behave as an OR on the overall activation reason
for the coil. This makes it like shots being repeated over the same switch,
but managed in different modes with different profiles exclusively.

Oh also added all of the disable events for the diverters to the deactivate
events because it seems like they remain diverting the full duration even if disabled.

This could also just be a virtual behavior. Wire it up!